Extract CommentForm component
- Cypress test is now passing as a result
